+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lite-ssr/core
 
+## 0.3.1
+
+### Patch Changes
+
+- Fix building index.html
+
 ## 0.3.0
 
 ### Minor Changes