+++ packages/lite-ssr/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.3.1
+
+### Patch Changes
+
+- Fix building index.html
+- Updated dependencies
+  - @lite-ssr/core@0.3.1
+
 ## 0.3.0
 
 ### Minor Changes